requirements.txt
flask~=3.0.3
PyMySQL~=1.1.1
python-dotenv~=1.0.1
SQLAlchemy~=2.0.31